requirements.txt
PyYAML
requests
datetime